Updated the deployment method to support multiple deployments
